requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
Babel==2.9.1
Beaker==1.10.0
beautifulsoup4==4.11.0
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
cloudinary==1.29.0
dj-database-url==1.0.0
Django==4.0.5
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-ckeditor==6.5.0
django-cloudinary-storage==0.3.0
django-dotenv==1.4.2
django-environ==0.9.0
django-js-asset==2.0.0
django-on-heroku==1.1.2
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
virtualenv==20.16.3
whitenoise==6.2.0
